Extract some steam logic to another module
I'm trying to add a bunch of steam processing functionality, so it's
easier if this is in another module.

Changed the scandir call to use pathlib paths instead of os.path since
steam_path is stored as a Path.
